+++ src/DonorProfiles/resources/js/app/tabs/donation-history/content.js
@@ -5,6 +5,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 const { __ } = wp.i18n;
 
 import Heading from '../../components/heading';
+import Button from '../../components/button';
 import DonationReceipt from '../../components/donation-receipt';
 import DonationTable from '../../components/donation-table';
 
@@ -42,13 +43,18 @@ const Content = () => {
 		) : (
 			<Fragment>
 				<Heading>
-					{ __( 'Donation', 'give' ) } #{ id }
+					{ __( 'Donation', 'give' ) } { getDonationById( id ).payment.serialCode }
 				</Heading>
 				<DonationReceipt donation={ getDonationById( id ) } />
-				<div className="give-donor-profile__donation-history-link">
+				<div className="give-donor-profile__donation-history-footer">
 					<Link to="/donation-history">
 						<FontAwesomeIcon icon="arrow-left" /> { __( 'Back to Donation History', 'give' ) }
 					</Link>
+					{ getDonationById( id ).payment.pdfReceiptUrl.length && (
+						<Button icon="file-pdf" href={ getDonationById( id ).payment.pdfReceiptUrl }>
+							{ __( 'Download Receipt', 'give' ) }
+						</Button>
+					) }
 				</div>
 			</Fragment>
 		);