+++ src/DonorProfiles/resources/js/app/tabs/edit-profile/content.js
@@ -7,6 +7,7 @@ import TextControl from '../../components/text-control';
 import RadioControl from '../../components/radio-control';
 import Button from '../../components/button';
 import { updateProfileWithAPI } from './utils';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 import EmailControls from './email-controls';
 import AddressControls from './address-controls';
@@ -20,6 +21,7 @@ import './style.scss';
 const Content = () => {
 	const id = useSelector( state => state.id );
 	const storedProfile = useSelector( state => state.profile );
+	const [ isUpdating, setIsUpdating ] = useState( false );
 
 	useEffect( () => {
 		setAvatarFile( null );
@@ -47,6 +49,8 @@ const Content = () => {
 	const [ firstName, setFirstName ] = useState( storedProfile.firstName );
 	const [ lastName, setLastName ] = useState( storedProfile.lastName );
 
+	const [ company, setCompany ] = useState( storedProfile.company );
+
 	const [ primaryEmail, setPrimaryEmail ] = useState( storedProfile.emails ? storedProfile.emails.primary : '' );
 
 	const reducedAdditionalEmails = storedProfile.emails ? Object.keys( storedProfile.emails ).reduce( ( newArray, key ) => {
@@ -68,30 +72,34 @@ const Content = () => {
 	}, [] ) : [];
 	const [ additionalAddresses, setAdditionalAddresses ] = useState( reducedAdditionalAddresses );
 
-	const [ anonymous, setAnonymous ] = useState( storedProfile.isAnonymous );
+	const [ isAnonymous, setIsAnonymous ] = useState( storedProfile.isAnonymous );
 	const anonymousOptions = [
 		{
-			value: 'public',
+			value: '0',
 			label: __( 'Public - show my donations publicly', 'give' ),
 		},
 		{
-			value: 'private',
+			value: '1',
 			label: __( 'Private - only organization admins can view my info' ),
 		},
 	];
 
-	const handleUpdate = () => {
-		updateProfileWithAPI( {
+	const handleUpdate = async() => {
+		setIsUpdating( true );
+		await updateProfileWithAPI( {
 			titlePrefix,
 			firstName,
 			lastName,
+			company,
 			primaryEmail,
 			additionalEmails,
 			primaryAddress,
 			additionalAddresses,
 			avatarFile,
+			isAnonymous,
 			id,
 		} );
+		setIsUpdating( false );
 	};
 
 	return (
@@ -127,6 +135,11 @@ const Content = () => {
 					onChange={ ( value ) => setLastName( value ) }
 				/>
 			</FieldRow>
+			<TextControl
+				label={ __( 'Company', 'give' ) }
+				value={ company }
+				onChange={ ( value ) => setCompany( value ) }
+			/>
 			<EmailControls
 				primaryEmail={ primaryEmail }
 				additionalEmails={ additionalEmails }
@@ -147,11 +160,12 @@ const Content = () => {
 				label={ __( 'Anonymous Giving' ) }
 				description={ __( 'This will prevent your avatar, first name, and donation comments and other information from appearing publicly on this orgization’s website.', 'give' ) }
 				options={ anonymousOptions }
-				onChange={ ( value ) => setAnonymous( value ) }
-				value={ anonymous }
+				onChange={ ( value ) => setIsAnonymous( value ) }
+				value={ isAnonymous }
 			/>
-			<Button icon="save" onClick={ () => handleUpdate() }>
-				Update Profile
+			<Button onClick={ () => handleUpdate() }>
+				{ __( 'Update Profile', 'give' ) }
+				<FontAwesomeIcon className={ isUpdating ? 'give-donor-profile__edit-profile-spinner' : '' } icon={ isUpdating ? 'spinner' : 'save' } fixedWidth />
 			</Button>
 		</Fragment>
 	);