+++ src/DonorProfiles/resources/js/app/tabs/edit-profile/utils/index.js
@@ -7,11 +7,13 @@ export const updateProfileWithAPI = async( {
 	titlePrefix,
 	firstName,
 	lastName,
+	company,
 	primaryEmail,
 	additionalEmails,
 	primaryAddress,
 	additionalAddresses,
 	avatarFile,
+	isAnonymous,
 	id,
 } ) => {
 	/**
@@ -28,16 +30,18 @@ export const updateProfileWithAPI = async( {
 	 * Pass new profile data to the Profile REST endpoint
 	 */
 	const { dispatch } = store;
-	axios.post( getAPIRoot() + 'give-api/v2/donor-profile/profile', {
+	return axios.post( getAPIRoot() + 'give-api/v2/donor-profile/profile', {
 		data: JSON.stringify( {
 			titlePrefix,
 			firstName,
 			lastName,
+			company,
 			primaryEmail,
 			additionalEmails,
 			primaryAddress,
 			additionalAddresses,
 			avatarId,
+			isAnonymous,
 		} ),
 		id,
 	}, {
@@ -52,6 +56,7 @@ export const updateProfileWithAPI = async( {
 			 * the donor's profile data
 			 */
 			dispatch( setProfile( responseData.profile ) );
+			return responseData;
 		} );
 };